+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_11_08_073028) do
+ActiveRecord::Schema.define(version: 2019_11_08_083442) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
@@ -277,6 +277,7 @@ ActiveRecord::Schema.define(version: 2019_11_08_073028) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "role", default: "user", null: false
+    t.string "remember_digest"
     t.index ["email"], name: "index_users_on_email", unique: true
   end